🏗️ `boostsrl.base.BaseBoostedRelationalModel` base class
Add `boostsrl/base.py` base class for deriving estimators
Add `boostsrl/tests/test_base.py` to test this base class
Refactor `boostsrl/rdn.py` to use the base class

Several estimators will need to be derived which need access
to a common set of methods. This base class handles the logic
for verifying parameters, checking estimators, and leaves
`fit()`, `predict()` and `predict_proba()` NotImplemented.

This is intended to help prevent code duplication as other
methods (Markov Logic Networks, RDNRegression, etc.) are
added.
