+++ boostsrl/rdn.py
@@ -5,21 +5,14 @@ Relational Dependency Networks
 """
 
 import re
-import subprocess
 import numpy as np
-from sklearn.base import BaseEstimator
-from sklearn.base import ClassifierMixin
-from sklearn.utils.validation import check_is_fitted
 
-from .background import Background
-from .system_manager import FileSystem
-from ._meta import DEBUG
-
+from .base import BaseBoostedRelationalModel
 
 # TODO: @property: feature_importances_
 
 
-class RDN(BaseEstimator, ClassifierMixin):
+class RDN(BaseBoostedRelationalModel):
     """Relational Dependency Networks Estimator
 
     Wrappers around BoostSRL for learning and inference with Relational Dependency
@@ -86,87 +79,14 @@ class RDN(BaseEstimator, ClassifierMixin):
         feature_importances_ : array, shape (n_features)
             Return the feature importances (based on how often each feature appears)
         """
-        self.background = background
-        self.target = target
-        self.n_estimators = n_estimators
-        self.node_size = node_size
-        self.max_tree_depth = max_tree_depth
-        self.debug = DEBUG
 
-        # Initialize the _FILE_SYSTEM to None. Replace if parameters are valid.
-        # self._FILE_SYSTEM = None
-        self.file_system = None
-
-    def _check_params(self):
-        """Check validity of parameters and raise ValueError if invalid.
-
-        If all parameters are valid, instantiate ``self.file_system`` by
-        instantiating it with a :class:`boostsrl.system_manager.FileSystem`
-        """
-        if self.target == "None":
-            raise ValueError("target must be set, cannot be {0}".format(self.target))
-        if not isinstance(self.target, str):
-            raise ValueError(
-                "target must be a string, cannot be {0}".format(self.target)
-            )
-        if self.background is None:
-            raise ValueError(
-                "background must be set, cannot be {0}".format(self.background)
-            )
-        if not isinstance(self.background, Background):
-            raise ValueError(
-                "background should be a boostsrl.Background object, cannot be {0}".format(
-                    self.background
-                )
-            )
-        if not isinstance(self.n_estimators, int) or isinstance(
-            self.n_estimators, bool
-        ):
-            raise ValueError(
-                "n_estimators must be an integer, cannot be {0}".format(
-                    self.n_estimators
-                )
-            )
-        if self.n_estimators <= 0:
-            raise ValueError(
-                "n_estimators must be greater than 0, cannot be {0}".format(
-                    self.n_estimators
-                )
-            )
-
-        # If all params are valid, allocate a FileSystem:
-        self.file_system = FileSystem()
-
-    def _check_initialized(self):
-        """Check for the estimator(s), raise an error if not found."""
-        check_is_fitted(self, "estimators_")
-
-    @staticmethod
-    def _call_shell_command(shell_command):
-        """Start a new process to execute a shell command.
-
-        This is intended for use in calling jar files. It opens a new process and
-        waits for it to return 0.
-
-        Parameters
-        ----------
-        shell_command : str
-            A string representing a shell command.
-
-        Returns
-        -------
-        None
-        """
-
-        # TODO: Explore other ways to interface with BoostSRL or the JVM.
-        #   https://wiki.python.org/moin/IntegratingPythonWithOtherLanguages#Java
-
-        _pid = subprocess.Popen(shell_command, shell=True)
-        _status = _pid.wait()
-        if _status != 0:
-            raise RuntimeError(
-                "Error when running shell command: {0}".format(shell_command)
-            )
+        super().__init__(
+            background=background,
+            target=target,
+            n_estimators=n_estimators,
+            node_size=node_size,
+            max_tree_depth=max_tree_depth,
+        )
 
     def fit(self, database):
         """Learn structure and parameters.
@@ -247,6 +167,7 @@ class RDN(BaseEstimator, ClassifierMixin):
         # TODO: On error, collect log files.
         return self
 
+    # TODO: This may be possible to set in the base class.
     def _run_inference(self, database) -> None:
         """Run inference mode on the BoostSRL Jar files.
